Update ex3.py
mudanca de int para float

+++ ex3.py
@@ -1,6 +1,6 @@
-dias = int(input("Informe os dias "))
-hrs = int(input("Informe as horas "))
-min = int(input("Informe os minutos "))
-sec = int(input("Informe os segundos "))
+dias = float(input("Informe os dias "))
+hrs = float(input("Informe as horas "))
+min = float(input("Informe os minutos "))
+sec = float(input("Informe os segundos "))
 
 print( "Total de", (dias*24*3600) + (hrs*3600) + (min*60) + sec , "segundos")
